Dedupe more byte-identical Integration/Database fixtures
