+++ tests/Database/Fixtures/Models/Attachment.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Attachment extends Model
+{
+    public $timestamps = false;
+}
