+++ tests/Database/Fixtures/Models/Guarded/Post.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures\Models\Guarded;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Post extends Model
+{
+    protected $guarded = [];
+}
