+++ tests/Database/Fixtures/Models/IntTimestampCasts/UnixTimeStampToCarbon.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures\Models\IntTimestampCasts;
+
+use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
+use Illuminate\Support\Carbon;
+
+class UnixTimeStampToCarbon implements CastsAttributes
+{
+    public function get($model, string $key, $value, array $attributes)
+    {
+        return Carbon::parse($value);
+    }
+
+    public function set($model, string $key, $value, array $attributes)
+    {
+        return Carbon::parse($value)->getTimestamp();
+    }
+}
