+++ tests/Database/Fixtures/Models/LoadAggregate/BaseModel.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures\Models\LoadAggregate;
+
+use Illuminate\Database\Eloquent\Model;
+
+class BaseModel extends Model
+{
+    public $timestamps = false;
+
+    protected $guarded = [];
+
+    public function related1()
+    {
+        return $this->hasMany(Related1::class);
+    }
+
+    public function related2()
+    {
+        return $this->hasMany(Related2::class);
+    }
+}
