+++ tests/Database/Fixtures/Models/MorphEagerLoading/User.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures\Models\MorphEagerLoading;
+
+use Illuminate\Database\Eloquent\Model;
+
+class User extends Model
+{
+    public $timestamps = false;
+}
