+++ tests/Database/Fixtures/Models/MorphEagerLoading/Video.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Illuminate\Tests\Database\Fixtures\Models\MorphEagerLoading;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Video extends Model
+{
+    public $timestamps = false;
+    protected $primaryKey = 'video_id';
+}
