+++ tests/Integration/Database/EloquentHasOneIsTest.php
@@ -5,6 +5,7 @@ namespace Illuminate\Tests\Integration\Database\EloquentHasOneIsTest;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Tests\Database\Fixtures\Models\Attachment;
 use Illuminate\Tests\Integration\Database\DatabaseTestCase;
 
 class EloquentHasOneIsTest extends DatabaseTestCase
@@ -82,11 +83,6 @@ class EloquentHasOneIsTest extends DatabaseTestCase
         $this->assertFalse($parent->attachment()->is($child));
         $this->assertTrue($parent->attachment()->isNot($child));
     }
-}
-
-class Attachment extends Model
-{
-    public $timestamps = false;
 }
 
 class Post extends Model
