+++ tests/Integration/Database/EloquentMorphCountLazyEagerLoadingTest.php
@@ -2,10 +2,11 @@
 
 namespace Illuminate\Tests\Integration\Database\EloquentMorphCountLazyEagerLoadingTest;
 
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Tests\Database\Fixtures\Models\Comment;
+use Illuminate\Tests\Database\Fixtures\Models\PostLikes\Like;
+use Illuminate\Tests\Database\Fixtures\Models\PostLikes\Post;
 use Illuminate\Tests\Integration\Database\DatabaseTestCase;
 
 class EloquentMorphCountLazyEagerLoadingTest extends DatabaseTestCase
@@ -45,25 +46,5 @@ class EloquentMorphCountLazyEagerLoadingTest extends DatabaseTestCase
 
         $this->assertTrue($comment->relationLoaded('commentable'));
         $this->assertEquals(2, $comment->commentable->likes_count);
-    }
-}
-
-class Post extends Model
-{
-    public $timestamps = false;
-
-    public function likes()
-    {
-        return $this->hasMany(Like::class);
-    }
-}
-
-class Like extends Model
-{
-    public $timestamps = false;
-
-    public function post()
-    {
-        return $this->belongsTo(Post::class);
     }
 }
