+++ tests/Integration/Database/EloquentMorphToIsTest.php
@@ -2,10 +2,10 @@
 
 namespace Illuminate\Tests\Integration\Database\EloquentMorphToIsTest;
 
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Tests\Database\Fixtures\Models\Comment;
+use Illuminate\Tests\Database\Fixtures\Models\MorphToTarget\Post;
 use Illuminate\Tests\Integration\Database\DatabaseTestCase;
 
 class EloquentMorphToIsTest extends DatabaseTestCase
@@ -84,9 +84,4 @@ class EloquentMorphToIsTest extends DatabaseTestCase
         $this->assertFalse($child->commentable()->is($parent));
         $this->assertTrue($child->commentable()->isNot($parent));
     }
-}
-
-class Post extends Model
-{
-    //
 }
