+++ tests/Integration/Database/EloquentPaginateTest.php
@@ -5,6 +5,7 @@ namespace Illuminate\Tests\Integration\Database;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Tests\Database\Fixtures\Models\Guarded\Post;
 
 class EloquentPaginateTest extends DatabaseTestCase
 {
@@ -96,11 +97,6 @@ class EloquentPaginateTest extends DatabaseTestCase
         $this->assertEquals(5, $query->count());
         $this->assertEquals(5, $query->paginate()->total());
     }
-}
-
-class Post extends Model
-{
-    protected $guarded = [];
 }
 
 class User extends Model
